requirements.txt
Django == 4.2.6
django-cors-headers
djangorestframework 
msal
azure-storage-blob
requests
psycopg2